refactor(uno): split logic/opponent/llm; host intents ledger + FLIP
- Extract uno types/helpers/logic/opponent/llm with thin index registerGameModule
- Wire multiplayer host onRemoteIntent through tryApplySoloSessionAction + flip capture
- gen-ai-heuristic-catalog: prefer opponent.ts *SelectAiAction JSDoc, fallback index selectAiAction
- Uno LLM hooks describeLegalChoice + summarizeLedgerAction
- AGENTS.md notes optional multi-file game layout
